requirements.txt
asgiref==3.2.10
Django==3.0.8
django-autoslug==1.9.7
django-ckeditor==5.9.0
django-js-asset==1.2.2
Paginator==0.5.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
